feat: 리더 사용자 탈퇴 차단 구현

+++ backend/src/main/java/withoutc/chongchong/user/service/UserService.java
@@ -3,8 +3,14 @@ package withoutc.chongchong.user.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import withoutc.chongchong.study.entity.StudyMember;
+import withoutc.chongchong.study.repository.StudyMemberRepository;
 import withoutc.chongchong.user.entity.User;
+import withoutc.chongchong.user.exception.UserErrorCode;
+import withoutc.chongchong.user.exception.UserException;
 import withoutc.chongchong.user.repository.UserRepository;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -12,10 +18,20 @@ import withoutc.chongchong.user.repository.UserRepository;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final StudyMemberRepository studyMemberRepository;
 
     @Transactional
     public void withdraw(Long userId) {
         User user = userRepository.getByIdForUpdateOrThrow(userId);
+        List<StudyMember> studyMembers = studyMemberRepository.findAllByUserIdOrderByCreatedAtDesc(userId);
+
+        boolean hasLeader = studyMembers.stream()
+                .anyMatch(StudyMember::isLeader);
+
+        if (hasLeader) {
+            throw new UserException(UserErrorCode.STUDY_LEADER_WITHDRAWAL_BLOCKED);
+        }
+
         userRepository.delete(user);
     }
 }
